+++ packages/pwa-kit-extension-support/package.json
@@ -40,6 +40,7 @@
     "@babel/core": "^7.21.3",
     "@babel/parser": "^7.21.3",
     "@types/express": "^5.0.0",
+    "@types/babel__core": "^7.20.5",
     "@types/fs-extra": "^11.0.4",
     "@types/hoist-non-react-statics": "~3.3.5",
     "@types/node": "^20",
