--- packages/pwa-kit-extension-support/src/configs/application-extensions.hbs
@@ -1,16 +1,0 @@
-/*
-* Copyright (c) 2024, salesforce.com, inc.
-* All rights reserved.
-* SPDX-License-Identifier: BSD-3-Clause
-* For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
-*/
-
-{{#each installed}}
-import {{getInstanceName .}} from {{.}}
-{{/each}}
-
-export default [
-    {{#each configured}}
-    new {{getInstanceName .}}({}){{#if (isNotLast @index @root.configured.length)}},{{/if}}
-    {{/each}}
-]
