+++ packages/pwa-kit-extension-support/src/configs/webpack/loaders/extensions-loader.ts
@@ -5,17 +5,12 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import dedent from 'dedent'
+// Third-Party
 import {LoaderContext} from 'webpack'
 import {PackageJson} from 'type-fest'
-import Handlebars from 'handlebars'
 
-import {kebabToUpperCamelCase} from '../../../shared/utils'
-import {nameRegex} from '../../../shared/utils/extensibility-utils'
-
-const APP_EXTENSION_CLIENT_ENTRY = 'setup-app'
-const APP_EXTENSION_SERVER_ENTRY = 'setup-server'
-const APP_EXTENSION_PREFIX = 'extension-' // aligns with what's in `nameRegex`
+// Local
+import {renderTemplate} from '../../utils'
 
 // TODO: Move these to a better location.
 interface ExtensionsLoaderOptions {
@@ -27,51 +22,6 @@ interface ExtensionsLoaderOptions {
 interface ExtensionsLoaderContext extends LoaderContext<ExtensionsLoaderOptions> {
     // You can add any additional properties if needed
 }
-
-const templateString = dedent`
-    /*
-    * Copyright (c) 2024, salesforce.com, inc.
-    * All rights reserved.
-    * SPDX-License-Identifier: BSD-3-Clause
-    * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
-    */
-
-    {{#each installed}}
-    import {{getInstanceName .}} from {{.}}
-    {{/each}}
-
-    export default [
-        {{#each configured}}
-        new {{getInstanceName .}}({}){{#if (isNotLast @index @root.configured.length)}},{{/if}}
-        {{/each}}
-    ]
-`
-const renderTemplate = (templateString: string, data: any) => {
-    const kebabToUpperCamelCase = (aString: string) => {
-        return aString.split('-')
-            .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
-            .join('')
-    }
-
-    Handlebars.registerHelper('getInstanceName', function (aString) {
-        const nameRegex = /^(?:@([^/]+)\/)?extension-(.+)$/
-        const [packageName, namespace, name] = aString.match(nameRegex)
-        
-        return kebabToUpperCamelCase(`${namespace ? `${namespace}-` : ''}-${name}`)
-    })
-      
-    Handlebars.registerHelper('isNotLast', function (index, arrayLength) {
-        return index !== arrayLength - 1;
-    })
-
-    // Compile the template
-    const template = Handlebars.compile(templateString)
-  
-    // Apply data to the compiled template
-    const output = template(data)
-  
-    return output
-  }
 
 /**
  * The `extensions-loader` as a mechanism to get all configured extensions for a given pwa-kit
@@ -87,90 +37,15 @@ const renderTemplate = (templateString: string, data: any) => {
  *
  * @returns {string} The string representation of a module exporting all the named application extension modules.
  */
-module.exports = function (this: ExtensionsLoaderContext, that: any) {
-    // console.log('this|that: ', this, that)
-    const {pkg, getConfig, target = 'web'} = that?.getOptions?.() || this.getOptions() || {}
-    const {devDependencies} = pkg
+module.exports = function (this: ExtensionsLoaderContext): string {
+    // const {pkg, getConfig, target = 'web'} = this.getOptions() || {}
+    const {target = 'node'} = this?.getOptions?.() || {}
 
-    // ------
-    console.log('process.cwd()', process.cwd())
-    const renderedHTML = renderTemplate(
-        templateString, 
+    return renderTemplate(
         {
-            installed: ['@salesforce/extension-sample', '@salesforce/extension-sample-2'],
+            installed: ['@salesforce/extension-sample'],
             configured: ['@salesforce/extension-sample'],
+            target
         }
     )
-    console.log(renderedHTML)
-    // ------
-    // TODO: clean this up.. looks like this is a bad variable name for the type that it represents.
-    const extensions = Object.keys(devDependencies || {})
-        .map((packageName) => packageName.match(nameRegex))
-        .filter(Boolean)
-
-    // Ensure that only valid extension names are loaded.
-    const extensionDetails = extensions.map((match) => {
-        const [packageName, namespace, name] = match || []
-        return {
-            instanceVariable: kebabToUpperCamelCase(`${namespace ? `${namespace}-` : ''}-${name}`),
-            modulePath: `${
-                namespace ? `@${namespace}/` : ''
-            }${APP_EXTENSION_PREFIX}${name}/${target === 'web' ? APP_EXTENSION_CLIENT_ENTRY : APP_EXTENSION_SERVER_ENTRY}`,
-            packageName
-        }
-    })
-    
-    const appExtensions = getConfig()?.app?.extensions
-
-    const result = dedent`
-            /*
-            * Copyright (c) 2024, salesforce.com, inc.
-            * All rights reserved.
-            * SPDX-License-Identifier: BSD-3-Clause
-            * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
-            */
-            // NOTE: I'm not completely sure why this was giving me isses as it should be running in the context of the
-            // base application and not the extension-support project. 
-            // import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
-
-            // TODO: Once we create @salesforce/pwa-kit-extensibility we will refactor this code to use a shared utility that
-            // normalizes/expands the configuration array before it is uses in this loader.
-            const normalizeExtensionsList = (extensions = []) =>
-                extensions.map((extension) => {
-                    return {
-                        packageName: Array.isArray(extension) ? extension[0] : extension,
-                        config: Array.isArray(extension) ? {enabled: true, ...extension[1]} : {enabled: true}
-                    }
-                })
-            
-            // App Extensions
-            ${extensionDetails
-                .map(
-                    ({instanceVariable, modulePath}) =>
-                        `import ${instanceVariable} from '${modulePath}'`
-                )
-                .join('\n            ')}
-
-            const installedExtensions = [
-                ${extensionDetails
-                    .map(
-                        ({packageName, instanceVariable}) =>
-                            `{packageName: '${packageName}', instanceVariable: ${instanceVariable}}`
-                    )
-                    .join(',\n                ')}
-            ]
-
-            const configuredExtensions = (normalizeExtensionsList(${JSON.stringify(appExtensions)}) || [])
-                .filter((extension) => extension.config.enabled)
-                .map((extension) => {
-                    // Make sure that the configured extensions are installed, before instantiating them
-                    const found = installedExtensions.find((ext) => ext.packageName === extension.packageName)
-                    return found ? new found.instanceVariable(extension.config || {}) : false
-                })
-                .filter(Boolean)
-
-            export default configuredExtensions
-        `
-    console.log('result: ', result)
-    return result
 }
